\name{anomalise}
\alias{anomalise}
\title{Create an anomaly time series from a RasterStack, either based on the
overall mean of the stack, or a supplied reference raster.}
\usage{
anomalise(x, reference = NULL, ...)
}
\description{
Create an anomaly time series from a RasterStack, either
based on the overall mean of the stack, or a supplied
reference raster.
}

