\name{denoise}
\alias{denoise}
\title{Noise filtering through principal components. The user can either specify
how many components to keep or can specify a minimum value for the variance that
should be kept.}
\usage{
denoise(data, k = NULL, expl.var = 0.95, weighted = TRUE, ...)
}
\description{
Noise filtering through principal components. The user can
either specify how many components to keep or can specify a
minimum value for the variance that should be kept.
}

