\name{EotCycle}
\alias{EotCycle}
\title{Calculates a single EOT and is controlled by the main eot() function}
\usage{
EotCycle(pred, resp, resp.eq.pred = F, n = 1, standardised, orig.var,
  write.out, path.out, names.out, type, print.console, ...)
}
\description{
Calculates a single EOT and is controlled by the main eot()
function
}

