% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{gimmsKiliNDVI}
\alias{gimmsKiliNDVI}
\title{Monthly GIMMS NDVI images for Mt. Kilimanjaro}
\format{a RasterBrick with the following attributes\cr
\cr
dimensions  : 8, 10, 80, 48  (nrow, ncol, ncell, nlayers)\cr
resolution  : 0.07272727, 0.07272727  (x, y)\cr
extent      : 37.02249, 37.74976, -3.431628, -2.84981  (xmin, xmax, ymin, ymax)\cr
coord. ref. : +proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs\cr}
\description{
Monthly Global Inventory Modelling and Mapping Studies (GIMMS)
NDVI images for the Mt. Kilimanjaro region from 2003/01 to 2006/12
}
\details{


Monthly Global Inventory Modelling and Mapping Studies (GIMMS)
NDVI images for the Mt. Kilimanjaro region from 2003/01 to 2006/12
}
\references{
An extended AVHRR 8 km NDVI dataset compatible with MODIS and SPOT vegetation NDVI data\cr
Tucker et al. (2005)\cr
International Journal of Remote Sensing, Volume 26, Isuue 20, pp. 4485 - 4498\cr
\url{http://www.tandfonline.com/doi/abs/10.1080/01431160500168686#.UuzaSx_PE24}
}

