% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{modisKiliNDVI}
\alias{modisKiliNDVI}
\title{Monthly MODIS NDVI images for Mt. Kilimanjaro}
\format{a RasterBrick with the following attributes\cr
\cr
dimensions  : 271, 345, 93495, 48  (nrow, ncol, ncell, nlayers)\cr
resolution  : 0.00215, 0.00215  (x, y)\cr
extent      : 36.98824, 37.72999, -3.411286, -2.828636  (xmin, xmax, ymin, ymax)\cr
coord. ref. : +proj=longlat +ellps=WGS84\cr}
\source{
\url{https://lpdaac.usgs.gov/}
}
\description{
Monthly MODIS Aqua MYD13Q1 NDVI images for the Mt. Kilimanjaro region from 2003/01 to 2006/12
}
\details{


Monthly MODIS Aqua MYD13Q1 NDVI images for the Mt. Kilimanjaro region from 2003/01 to 2006/12
}
\references{
MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid, MYD13Q1 V005.\cr
NASA Land Processes Distributed Active Archive Center (LP DAAC) 2006\cr
Sioux Falls, South Dakota, USA\cr
}

