\name{plotEot}
\alias{plotEot}
\title{Standard plotting routine for the results of eot().}
\usage{
plotEot(eot.obj, eot = 1, pred.prm = "rsq.predictor",
  resp.prm = "rsq.response", show.eot.loc = FALSE, anomalies = TRUE,
  add.map = TRUE, times.vec = NULL, arrange = c("wide", "long"),
  clr = colorRampPalette(rev(brewer.pal(9, "Spectral")))(1000), ...)
}
\description{
Standard plotting routine for the results of eot().
}

