\name{plotLocations}
\alias{plotLocations}
\title{A simple plotting routine to visualise the location of all identified base
points colour coded according to eot mode (1 - k).}
\usage{
plotLocations(eot.obj, ...)
}
\description{
A simple plotting routine to visualise the location of all
identified base points colour coded according to eot mode
(1 - k).
}

