\name{Reot}
\alias{Reot}
\alias{Reot}
\docType{package}
\title{
Empirical Orthogonal Teleconnections in R
}
\description{
Empirical Orthogonal Teleconnections in R
}
\details{
\tabular{ll}{
Package: \tab Reot\cr
Type: \tab Package\cr
Version: \tab Version: 0.0.1\cr
Date: \tab 2014-01-31\cr
License: \tab GNU GPL (>=3)\cr
}
A collection of functions to facilitate empirical orthogonal teleconnection analysis. Some handy functions for preprocessing, such as deseasoning, denoising, lagging are readily available for ease of usage.
}
\author{
Tim Appelhans, Florian Detsch

Maintainer: Tim Appelhans <tim.appelhans@gmail.com>
}
\references{
Empirical Orthogonal Teleconnections\cr
H. M. van den Dool, S. Saha, Å. Johansson (2000)\cr
Journal of Climate, Volume 13, Issue 8 (April 2000) pp. 1421-1435\cr

Empirical methods in short-term climate prediction\cr
H. M. van den Dool (2007)\cr
Oxford University Press, Oxford, New York (2007)\cr

}

\keyword{ package }
\seealso{

}
\examples{
}
