\name{resizingTextGrob}
\alias{resizingTextGrob}
\title{Some helper functions especially for the grid layout (from
http://ryouready.wordpress.com/2012/08/01/).}
\usage{
resizingTextGrob(..., scale.fact = 1)
}
\description{
Some helper functions especially for the grid layout (from
http://ryouready.wordpress.com/2012/08/01/).
}

