\name{nEot4Var}
\alias{nEot4Var}
\title{Function to identify the number of modes needed to explain a certain amount of
variance within the response series.}
\usage{
nEot4Var(eot.obj, var = 0.9)
}
\description{
Function to identify the number of modes needed to explain
a certain amount of variance within the response series.
}

