% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Reot-package}
\alias{Reot}
\alias{Reot-package}
\title{Empirical Orthogonal Teleconnections in R}
\description{
Empirical Orthogonal Teleconnections in R
}
\details{
A collection of functions to facilitate empirical orthogonal teleconnection analysis.
Some handy functions for preprocessing, such as deseasoning, denoising, lagging
are readily available for ease of usage.
}
\author{
Tim Appelhans, Florian Detsch\cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\references{
Empirical Orthogonal Teleconnections\cr
H. M. van den Dool, S. Saha, A. Johansson (2000)\cr
Journal of Climate, Volume 13, Issue 8 (April 2000) pp. 1421 - 1435\cr

Empirical methods in short-term climate prediction\cr
H. M. van den Dool (2007)\cr
Oxford University Press, Oxford, New York (2007)\cr
}
\seealso{
\pkg{Reot} is built upon raster* classes from the \code{\link{raster-package}}.
Please see their documentation for data preparation etc.
}
\keyword{package}

